Refactor: Atom 컴포넌트 구조 변경
