+++ src/components/atoms/SafeImage/index.ts
@@ -1,0 +1,1 @@
+export * from './SafeImage';
