+++ src/components/post/PostCommentForm.tsx
@@ -4,7 +4,7 @@ import { ArrowUp, MessageCircle } from 'lucide-react';
 
 import { FormEvent, useEffect, useState } from 'react';
 
-import Button from '@/components/common/Button';
+import { Button } from '@/components/atoms/Button';
 import Input from '@/components/common/Input';
 import { EventName } from '@/lib/firebase/eventNames';
 import { firebaseLogging } from '@/lib/firebase/logEvent';
