+++ src/stories/Sidebar.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Button from '@/components/common/Button';
+import { Button } from '@/components/atoms/Button';
 import Sidebar from '@/components/common/Sidebar';
 import ToastContainer from '@/components/common/ToastContainer';
 import { useSidebarStore } from '@/stores/sidebarStore';
